Change to fixed number of terms, not order
Previously the 'order' of a Fourier series meant that it was using twice
as many basis vectors in its series as the other two, so appeared to be
significantly more accurate.  Swap to a fixed number of terms, with
Fourier gaining a single odd or even term at each step, rather than one
of each.